Revert "[UI] 매거진 페이지 UI 구현"
