+++ .github/ISSUE_TEMPLATE/custom.md
@@ -1,0 +1,16 @@
+---
+name: Custom issue template
+about: Describe this issue template's purpose here.
+title: "[Label] 작업 내용"
+labels: ''
+assignees: ''
+
+---
+
+# 📋 설명
+- 작업하려는 이슈에 대해 설명해주세요
+
+# ✅ 체크리스트
+- [ ] TODO
+- [ ] TODO
+- [ ] TODO
