+++ .github/pull_request_template.md
@@ -1,0 +1,8 @@
+# 📌 이슈번호
+- 이슈번호를 적어주세요.
+
+# 📌 구현/추가 사항
+- 구현/추가 사항, 코멘트가 있다면 적어주세요.
+
+# 📷 미리보기
+- 구현한 내용을 보여주세요.
