+++ HMOA_iOS/HMOA_iOS/Resource/Assets.xcassets/HPedia/selectedAddButton.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "writeButtom.svg",
+      "filename" : "글쓰기 눌렀을 때 버튼.svg",
       "idiom" : "universal",
       "scale" : "1x"
     },
