--- HMOA_iOS/HMOA_iOS/Resource/Assets.xcassets/Magazine/Contents.json
@@ -1,6 +1,0 @@
-{
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
-}
