+++ HMOA_iOS/HMOA_iOS/Source/Extenstions/UIColor++Extenstions.swift
@@ -62,8 +62,3 @@ extension UIColor {
     }
 }
 
-extension UIColor {
-    static var random: UIColor {
-        UIColor(red: .random(in: 0...1), green: .random(in: 0...1), blue: .random(in: 0...1), alpha: 1.0)
-    }
-}
