+++ HMOA_iOS/HMOA_iOS/Source/Extenstions/UIViewController++Extenstions.swift
@@ -49,7 +49,7 @@ extension UIViewController {
             self.view.window?.rootViewController?.dismiss(animated: true)
         }
     }
-    
+        
     /// community Image DetailVC로 전환
     func presentImageListVC(_ indexPath: IndexPath, images: [CommunityPhoto]) {
         let vc = ImageListViewController()
@@ -74,7 +74,7 @@ extension UIViewController {
         CommunityDetailVC.hidesBottomBarWhenPushed = true
         self.navigationController?.pushViewController(CommunityDetailVC, animated: true)
     }
-    
+
     /// hpediaHomeVC, writedPostVC -> communityDetailVC
     func presentCommunityDetailVC(_ id: Int) {
         let CommunityDetailVC = CommunityDetailViewController()
@@ -229,13 +229,6 @@ extension UIViewController {
         self.navigationController?.pushViewController(totalPerfumeVC, animated: true)
     }
     
-    /// MagazineDetailVC로 push
-    func presentMagazineDetailViewController() {
-        let magazineDetailVC = MagazineDetailViewController()
-        magazineDetailVC.reactor = MagazineDetailReactor()
-        self.navigationController?.pushViewController(magazineDetailVC, animated: true)
-    }
-    
     // MARK: Configure NavigationBar
     
     /// 확인 버튼, 취소 버튼 navigation bar
@@ -306,30 +299,6 @@ extension UIViewController {
         self.navigationItem.rightBarButtonItems = [homeButton]
     }
     
-    /// Back 버튼, Share 버튼 NavigationBar
-    func setBackShareRightNaviBar(_ title: String) {
-        let titleLabel = UILabel().then {
-            $0.text = title
-            $0.font = .customFont(.pretendard_medium, 20)
-            $0.textColor = .black
-        }
-        
-        let backButton = self.navigationItem.makeImageButtonItem(self, action: #selector(popViewController), imageName: "backButton")
-        
-        let shareButton = self.navigationItem.makeImageButtonItem(self, action: #selector(shareMagazine), imageName: "share")
-
-        let appearance = UINavigationBarAppearance()
-        appearance.backgroundColor = .white
-        
-        self.navigationController?.navigationBar.standardAppearance = appearance
-        self.navigationController?.navigationBar.compactAppearance = appearance
-        self.navigationController?.navigationBar.scrollEdgeAppearance = appearance
-        
-        self.navigationItem.titleView = titleLabel
-        self.navigationItem.leftBarButtonItems = [backButton]
-        self.navigationItem.rightBarButtonItems = [shareButton]
-    }
-    
     /// NavigationBarTitle 설정
     func setNavigationBarTitle(_ title: String) {
         let titleLabel = UILabel().then {
@@ -386,10 +355,6 @@ extension UIViewController {
     /// brandSearchViewController로 push
     @objc func goToBrand() {
         presentBrandSearchViewController()
-    }
-    
-    @objc func shareMagazine() {
-        
     }
     
     // MARK: - UI Function
