+++ HMOA_iOS/HMOA_iOS/Source/Network/API/Home/Model/HomeData.swift
@@ -27,3 +27,6 @@ struct RecommendPerfume: Codable, Hashable{
 
 
 typealias HomeSecondData = [RecommendPerfumeList]
+
+
+
