--- HMOA_iOS/HMOA_iOS/Source/Presentation/Comon/View/SupplementaryViewKind.swift
@@ -1,11 +1,0 @@
-//
-//  SupplementaryViewKind.swift
-//  HMOA_iOS
-//
-//  Created by 곽다은 on 3/17/24.
-//
-
-enum SupplementaryViewKind {
-    static let header = "header"
-    static let background = "background"
-}
