+++ README.md
@@ -1,0 +1,24 @@
+# 향모아 HMOA
+내 손안에 향기를 담다<br/>
+생생한 향수 정보 제공 어플 HMOA
+
+## ✅ 주요 기능
+<img src="https://github.com/HMOAA/HMOA_iOS/assets/75370733/23e966b4-1748-45d8-98ac-3539fb3bd37e"  width="200" height="400"/>
+<img src="https://github.com/HMOAA/HMOA_iOS/assets/75370733/50fd6139-5757-4b49-bc60-826fd61be197"  width="200" height="400"/>
+<img src="https://github.com/HMOAA/HMOA_iOS/assets/75370733/7fdbd5cc-c22e-4ee6-8c5a-5bf862e063d1"  width="200" height="400"/>
+<img src="https://github.com/HMOAA/HMOA_iOS/assets/75370733/558d285f-51e8-4ff7-8dc2-0aa847128d12"  width="200" height="400"/>
+<img src="https://github.com/HMOAA/HMOA_iOS/assets/75370733/d74c4a1c-4814-4e32-9883-df96086987a3"  width="200" height="400"/>
+<img src="https://github.com/HMOAA/HMOA_iOS/assets/75370733/9c6eadec-39bf-4f42-9d86-ae9f6e16a003"  width="200" height="400"/>
+
+## ⚒️ 기술 스택
+<div>
+  <img src="https://img.shields.io/badge/UIKit-2396F3?style=flat&logo=UIKit&logoColor=white" />
+  <img src="https://img.shields.io/badge/Swift-F05138?style=flat&logo=Swift&logoColor=white" />
+  <img src="https://img.shields.io/badge/RxSwift-B7178C?style=flat&logo=RxSwift&logoColor=white" />
+  <img src="https://img.shields.io/badge/ReactorKit-0088CC?style=flat&logo=ReactorKit&logoColor=white" />
+  <img src="https://img.shields.io/badge/SnapKit-83B81Astyle=flat&logo=SnapKit&logoColor=white" />
+</div>
+
+## 📲 아키텍처
+- ReactorKit을 활용해 자연스러운 코드 컨벤션을 통일 시켰습니다.
+- SnapKit을 활용한 CodeBase로 구현했습니다.
